fix tooltip lag when changing hovered series
