+++ frontend/src/metabase/visualizations/echarts/cartesian/waterfall/option/index.ts
@@ -256,7 +256,6 @@ export const getWaterfallChartOption = (
       chartMeasurements,
       settings,
       hasTimelineEvents,
-      null,
       renderingContext,
     ),
     tooltip: getTooltipOption(chartModel, settings, false, false),
