+++ frontend/src/metabase/visualizations/visualizations/CartesianChart/use-chart-events.ts
@@ -30,7 +30,10 @@ import {
 } from "metabase/visualizations/visualizations/CartesianChart/events";
 import type { CardId } from "metabase-types/api";
 
-import { getHoveredEChartsSeriesDataKeyAndIndex } from "./utils";
+import {
+  getHoveredEChartsSeriesDataKeyAndIndex,
+  getHoveredSeriesDataKey,
+} from "./utils";
 
 export const useChartEvents = (
   chartRef: React.MutableRefObject<EChartsType | undefined>,
@@ -67,6 +70,42 @@ export const useChartEvents = (
       }
     },
     [card, onChangeCardAndRun, rawSeries],
+  );
+
+  const hoveredSeriesDataKey = useMemo(
+    () => getHoveredSeriesDataKey(chartModel.seriesModels, hovered),
+    [chartModel.seriesModels, hovered],
+  );
+
+  useEffect(
+    function updateYAxisVisibility() {
+      const hasSingleYAxis = !(
+        chartModel.leftAxisModel != null && chartModel.rightAxisModel != null
+      );
+
+      if (hasSingleYAxis) {
+        return;
+      }
+
+      const yAxisShowOption = [{ show: true }, { show: true }];
+      if (hoveredSeriesDataKey != null) {
+        const hiddenYAxisIndex = chartModel.leftAxisModel?.seriesKeys.includes(
+          hoveredSeriesDataKey,
+        )
+          ? 0
+          : 1;
+
+        yAxisShowOption[hiddenYAxisIndex].show = false;
+      }
+
+      chartRef.current?.setOption({ yAxis: yAxisShowOption }, false, true);
+    },
+    [
+      chartModel.leftAxisModel,
+      chartModel.rightAxisModel,
+      chartRef,
+      hoveredSeriesDataKey,
+    ],
   );
 
   const eventHandlers: EChartsEventHandler[] = useMemo(
